changed db to float

+++ db/migrate/20190528173046_create_scooters.rb
@@ -2,8 +2,8 @@ class CreateScooters < ActiveRecord::Migration[5.2]
   def change
     create_table :scooters do |t|
       t.string :company
-      t.decimal :latitude
-      t.decimal :longitude
+      t.float :latitude
+      t.float :longitude
       t.string :battery_level
 
       t.timestamps
